# openapi.yaml
openapi: 3.1.0
info:
  title: Afere API
  version: 3.1.0
  description: >
    Composition-builder API for the Afere medical procedure valuation platform.
    Implements CBHPM 2022 items 4.1, 4.2, and 5.1 for correct multi-procedure
    discounting and auxiliary surgeon remuneration.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/calculate:
    post:
      summary: Calculate total valuation for a composition
      description: >
        Accepts a physician-assembled composition of CBHPM codes plus staffing
        configuration and access route type, then returns a detailed breakdown
        implementing CBHPM 2022 items 4.1/4.2 (multi-procedure discounting) and
        item 5.1 (auxiliary surgeon remuneration at 60/40/30/30%).
      operationId: calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRequest'
      responses:
        '200':
          description: Detailed valuation breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponse'
        '400':
          description: Invalid request body.
  /api/compositions:
    post:
      summary: Save a surgical composition
      description: >
        Persists a reusable surgical template containing the physician's procedural
        setup. No financial values are stored — values are always recalculated fresh
        when the composition is executed.
      operationId: saveComposition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCompositionRequest'
      responses:
        '201':
          description: Composition saved. Returns the public_id for URL construction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveCompositionResponse'
        '400':
          description: Validation error (missing or invalid fields).
    get:
      summary: List saved compositions
      operationId: listCompositions
      responses:
        '200':
          description: Ordered list of compositions (newest first, up to 100).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompositionItem'
  /api/compositions/{id}:
    get:
      summary: Get a composition by public ID
      operationId: getComposition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full composition detail including selected codes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompositionDetail'
        '404':
          description: Composition not found.
    put:
      summary: Update an existing composition
      operationId: updateComposition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCompositionRequest'
      responses:
        '200':
          description: Updated composition detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompositionDetail'
        '404':
          description: Composition not found.
    delete:
      summary: Delete a composition
      operationId: deleteComposition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Composition deleted.
        '404':
          description: Composition not found.
...
components:
  schemas:
...
    AccessRouteType:
      type: string
      enum: [same, different]
      description: >
        Whether all selected procedures were performed via the same access route
        (CBHPM 4.1 – 50% discount on secondary procedures) or different access
        routes (CBHPM 4.2 – 70% discount on secondary procedures).
    SelectedCode:
      type: object
      required: [cbhpm_code, description, porte]
      properties:
        cbhpm_code:
          type: string
          description: CBHPM code selected by the physician.
        description:
          type: string
          description: CBHPM description (passed through from the procedure detail).
        porte:
          type: string
          description: Porte for this code (from the SBN catalog, not editable by the physician).
    CalculateRequest:
      type: object
      required: [selected_codes, auxiliaries_count, requires_anesthesia, access_route_type]
      properties:
        selected_codes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SelectedCode'
          description: Physician-selected CBHPM codes.
        auxiliaries_count:
          type: integer
          minimum: 0
          maximum: 4
          description: Number of auxiliary surgeons (0–4).
        requires_anesthesia:
          type: boolean
          description: Whether an anesthesiologist is required.
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        urgency_emergency:
          type: boolean
          description: >
            When true, applies the CBHPM item 2 urgency/emergency 30% surcharge
            to all medical fees (surgeon, auxiliaries, anesthesiologist).
            Applicable for acts performed 19h–7h, on weekends, or on holidays.
    CodeBreakdown:
      type: object
      required: [cbhpm_code, description, porte, base_value, is_principal]
      properties:
        cbhpm_code:
          type: string
        description:
          type: string
        porte:
          type: string
        base_value:
          type: number
          format: float
          description: Full monetary value of this code's porte (before any discount).
        is_principal:
          type: boolean
          description: True for the highest-value code selected as the principal procedure.
    AuxiliaryFee:
      type: object
      required: [position, percentage, fee]
      properties:
        position:
          type: integer
          description: Auxiliary position (1 = first, 2 = second, etc.).
        percentage:
          type: number
          format: float
          description: Percentage applied (60, 40, 30, or 30).
        fee:
          type: number
          format: float
          description: Calculated fee for this auxiliary.
    SurgeonBreakdown:
      type: object
      required: [principal_value, additional_gross, discount_rate, additional_discounted, surgeon_total]
      properties:
        principal_value:
          type: number
          format: float
          description: Monetary value of the principal (highest-value) procedure.
        additional_gross:
          type: number
          format: float
          description: Sum of all non-principal procedure values before applying the CBHPM discount.
        discount_rate:
          type: number
          format: float
          description: Discount rate applied to additional procedures (0.50 for same route, 0.70 for different routes).
        additional_discounted:
          type: number
          format: float
          description: Sum of additional procedure values after applying the discount.
        surgeon_total:
          type: number
          format: float
          description: Final surgeon fee (principal_value + additional_discounted).
    CalculateResponse:
      type: object
      required:
        - code_breakdown
        - access_route_type
        - surgeon_breakdown
        - lead_surgeon_fee
        - individual_auxiliary_fees
        - auxiliaries_fee
        - anesthesiologist_fee
        - final_total
        - total_base
      properties:
        code_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/CodeBreakdown'
          description: Per-code contribution with principal flag.
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        surgeon_breakdown:
          $ref: '#/components/schemas/SurgeonBreakdown'
        lead_surgeon_fee:
          type: number
          format: float
          description: Final surgeon fee (equals surgeon_breakdown.surgeon_total).
        individual_auxiliary_fees:
          type: array
          items:
            $ref: '#/components/schemas/AuxiliaryFee'
          description: Per-auxiliary fee details (CBHPM 5.1 – 60/40/30/30%).
        auxiliaries_fee:
          type: number
          format: float
          description: Sum of all individual auxiliary fees.
        anesthesiologist_fee:
          type: number
          format: float
          description: Fixed anesthesiologist fee (if applicable).
        final_total:
          type: number
          format: float
          description: Grand total (lead_surgeon_fee + auxiliaries_fee + anesthesiologist_fee).
        total_base:
          type: number
          format: float
          description: Simple sum of all selected code porte values (before discounting). Retained for compatibility.
        urgency_emergency_applied:
          type: boolean
          description: Whether the CBHPM item 2 urgency/emergency surcharge was applied.
        urgency_emergency_percentage:
          type: number
          format: float
          description: Surcharge percentage (30.0 when applied, 0 otherwise).
        urgency_emergency_value:
          type: number
          format: float
          description: Absolute monetary value of the surcharge added to the total.
    SaveCompositionRequest:
      type: object
      required: [name, sbn_procedure_name, selected_codes, auxiliaries_count, requires_anesthesia, access_route_type]
      properties:
        name:
          type: string
          minLength: 1
          description: Physician-defined label for this composition.
        sbn_procedure_id:
          type: string
          description: SBN catalog ID of the primary procedure (optional).
        sbn_procedure_name:
          type: string
          description: SBN procedure name as displayed to the physician.
        selected_codes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SelectedCode'
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        auxiliaries_count:
          type: integer
          minimum: 0
          maximum: 4
        requires_anesthesia:
          type: boolean
        urgency_emergency:
          type: boolean
          description: Persists the CBHPM item 2 urgency/emergency surcharge preference.
    SaveCompositionResponse:
      type: object
      required: [public_id, created_at]
      properties:
        public_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    CompositionItem:
      type: object
      required: [public_id, name, sbn_procedure_name, access_route_type, auxiliaries_count, requires_anesthesia, created_at]
      description: Lightweight composition projection for list responses (no selected codes).
      properties:
        public_id:
          type: string
          format: uuid
        name:
          type: string
        sbn_procedure_id:
          type: string
        sbn_procedure_name:
          type: string
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        auxiliaries_count:
          type: integer
        requires_anesthesia:
          type: boolean
        created_at:
          type: string
          format: date-time
    CompositionDetail:
      type: object
      required: [public_id, name, sbn_procedure_name, selected_codes, access_route_type, auxiliaries_count, requires_anesthesia, urgency_emergency, created_at, updated_at]
      description: Full composition including selected CBHPM codes.
      properties:
        public_id:
          type: string
          format: uuid
        name:
          type: string
        sbn_procedure_id:
          type: string
        sbn_procedure_name:
          type: string
        selected_codes:
          type: array
          items:
            $ref: '#/components/schemas/SelectedCode'
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        auxiliaries_count:
          type: integer
        requires_anesthesia:
          type: boolean
        urgency_emergency:
          type: boolean
          description: Whether the CBHPM item 2 urgency/emergency surcharge is active.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...